DRY up the test organization
