+++ .github/workflows/remoteconfig.yml
@@ -31,8 +31,8 @@ jobs:
           FirebaseRemoteConfig/Tests/SwiftAPI/GoogleService-Info.plist "$plist_secret"
     - name: BuildAndUnitTest # can be replaced with pod lib lint with CocoaPods 1.10
       run: scripts/third_party/travis/retry.sh scripts/build.sh RemoteConfig ${{ matrix.target }} unit
-    - name: Hermetic API Tests
-      run: scripts/third_party/travis/retry.sh scripts/build.sh RemoteConfig iOS hermetic
+    - name: Fake Console API Tests
+      run: scripts/third_party/travis/retry.sh scripts/build.sh RemoteConfig iOS fakeconsole
     - name: IntegrationTest
       if: matrix.target == 'iOS'
       run: ([ -z $plist_secret ] || scripts/third_party/travis/retry.sh scripts/build.sh RemoteConfig iOS integration)
