+++ FirebaseRemoteConfig/Tests/FakeUtils/GoogleService-Info.plist
@@ -19,7 +19,7 @@
 	<key>BUNDLE_ID</key>
 	<string>org.cocoapods.AppHost-FirebaseRemoteConfig-Unit-Tests</string>
 	<key>PROJECT_ID</key>
-	<string>abc-xyz-123</string>
+	<string>FakeProject</string>
 	<key>DATABASE_URL</key>
 	<string>https://abc-xyz-123.firebaseio.com</string>
 	<key>STORAGE_BUCKET</key>
